move FileUploadDialog handlers and state to hook
